feat: evaluator done

+++ crates/common/src/datalake/base.rs
@@ -7,7 +7,7 @@ type DataCompiler = dyn Fn() -> Result<Vec<DataPoint>>;
 /// DataPoint is a type that can be used to store data in a Datalake
 #[derive(Debug, Clone, PartialEq, PartialOrd)]
 pub enum DataPoint {
-    Int(usize),
+    Int(u64),
     Float(f64),
     Str(String),
 }
